+++ common/src/main/java/net/mehvahdjukaar/supplementaries/common/block/IKeyLockable.java
@@ -33,7 +33,8 @@ public interface IKeyLockable {
 
 
     default boolean shouldShowPassword() {
-        return this.getPassword().length() <= MAX_ITEM_NAME_LEN;
+        String password = this.getPassword();
+        return password != null && password.length() <= MAX_ITEM_NAME_LEN;
     }
 
     default void onKeyAssigned(Level level, BlockPos pos, Player player, String newKey) {